snap-env: add the ability to specify kilo/mega/giga
This commit adds the ability to specify a numeric value
in kilo/mega/giga bytes (capital characters supported as well).

e.g.
 env MEM_POOL_SIZE=1g amples/mlnx_snap_emu -m 0xff -u --mem-size 1200
 env MEM_POOL_SIZE=5G amples/mlnx_snap_emu -m 0xff -u --mem-size 1200

+++ src/snap_env.h
@@ -58,7 +58,7 @@ __attribute__((constructor)) void snap_env_register_##env_name(void) \
  *
  * Returns: -EINVAL on failure, integer value >= 0 upon success
  */
-int snap_env_getenv(const char *env_name);
+long long snap_env_getenv(const char *env_name);
 
 /**
  * snap_env_is_set - check if the env. variable was set
